+++ libmicrocosm/stat/fstat.c
@@ -1,9 +1,10 @@
 #include <sys/stat.h>
+#include <sys/types.h>
 
 #include "reerrno.h"
 #include "vfs/vfs.h"
 
-long MC_fstat(int fd, struct MC_struct_stat *buf)
+ssize_t MC_fstat(int fd, struct MC_struct_stat *buf)
 {
     int ret;
     struct stat hbuf;
