+++ libmicrocosm/unistd/getgid.c
@@ -1,2 +1,4 @@
+#include <sys/types.h>
+
 /* our IDs are fake */
-long MC_getgid() { return 1; }
+ssize_t MC_getgid() { return 1; }
