+++ libmicrocosm/unistd/setgid.c
@@ -1,7 +1,9 @@
+#include <sys/types.h>
+
 /* our IDs are fake */
 #include "mcerrno.h"
 
-long MC_setgid(long id) {
+ssize_t MC_setgid(ssize_t id) {
     if (id == 1) return 0;
     return -MC_EPERM;
 }
